+++ Game/CameraMovement.cpp
@@ -1,4 +1,6 @@
 #include "CameraMovement.h"
+#include "Core/Engine.h"
+#include "Core/Window.h"
 #include "Scene/Node.h"
 
 void CameraMovement::Run(Node* node) {
@@ -14,8 +16,8 @@ void CameraMovement::Update(Node* node) {
 	const Window* window = GetWindow();
 
 	float rotateSpeed = 0.3f;
-	static float lastX = window->m_mouse.GetX();
-	static float lastY = window->m_mouse.GetX();
+	static float lastX = (float) window->m_mouse.GetX();
+	static float lastY = (float) window->m_mouse.GetX();
 	if (window->m_mouse.IsLeftPressed()) {
 		float x = window->m_mouse.GetX();
 		float y = window->m_mouse.GetY();
